+++ README SOLANA.md
@@ -7,6 +7,6 @@
 3. Haven't tested with multiple items, so make sure in My Items/Owned you have no items. You can put existing items on sale.
 4. Create your item with Create button
 | --------------------------------------- | 
-5. Set SOL_PUBLIC_KEY and SOL_SECRET_KEY in env.devnet file
+5. Set SECRET_KEY_SOL in env.devnet file
 6. Then go to root folder of the SDK, run `NODE_ENV=devnet npx ts-node examples/solana/listForSale.ts`
 7. You can try to buy the item by switching wallet
