+++ src/interfaces.ts
@@ -1,0 +1,4 @@
+export interface RefinableOptions {
+    waitConfirmations?: number;
+    apiUrl?: string;
+}
