+++ src/interfaces/Network.ts
@@ -1,3 +1,6 @@
+import { ENV as ChainId } from '@solana/spl-token-registry';
+import { clusterApiUrl } from '@solana/web3.js';
+
 export enum ChainType {
   BSC = "bsc",
   POLYGON = "polygon",
@@ -20,3 +23,40 @@ export enum Chain {
   Ethereum = 1,
   EthereumRinkeby = 4,
 }
+
+export type ENV_SOL =
+  | 'mainnet-beta'
+  | 'mainnet-beta (Solana)'
+  | 'mainnet-beta (Serum)'
+  | 'testnet'
+  | 'devnet'
+  | 'localnet'
+  | 'lending';
+
+export const ENDPOINTS_SOL = [
+  {
+    name: 'mainnet-beta' as ENV_SOL,
+    endpoint: 'https://api.metaplex.solana.com/',
+    ChainId: ChainId.MainnetBeta,
+  },
+  {
+    name: 'mainnet-beta (Solana)' as ENV_SOL,
+    endpoint: 'https://api.mainnet-beta.solana.com',
+    ChainId: ChainId.MainnetBeta,
+  },
+  {
+    name: 'mainnet-beta (Serum)' as ENV_SOL,
+    endpoint: 'https://solana-api.projectserum.com/',
+    ChainId: ChainId.MainnetBeta,
+  },
+  {
+    name: 'testnet' as ENV_SOL,
+    endpoint: clusterApiUrl('testnet'),
+    ChainId: ChainId.Testnet,
+  },
+  {
+    name: 'devnet' as ENV_SOL,
+    endpoint: clusterApiUrl('devnet'),
+    ChainId: ChainId.Devnet,
+  },
+];
